# api/src/main/resources/openapi/api.yaml
openapi: 3.1.0
info:
  title: Financial Calculator API
  version: 1.0.0
  description: API for Financial Calculator Platform

servers:
  - url: http://localhost:8080
    description: Local DEV server

security:
  - cookieAuth: []

paths:
  # AUTH
  /auth/login:
    post:
      summary: User Login
      operationId: login
      tags: [Auth]
      security: [] # This endpoint is public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful. Returns user info in body and tokens in HttpOnly cookies.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: Invalid credentials

  /auth/refresh:
    post:
      summary: Refresh access token
      operationId: refreshToken
      tags: [Auth]
      responses:
        '200':
          description: Token refreshed successfully. Returns new user info and sets new access token cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: Invalid or expired refresh token

  /auth/logout:
    post:
      summary: User Logout
      operationId: logout
      tags: [Auth]
      responses:
        '204':
          description: Logout successful. Clears cookies.

  /auth/register:
    post:
      summary: Register new user
      operationId: register
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '409':
          description: Username or email already in use.

  # USERS
  /api/users:
    get:
      summary: Get all users
      operationId: getAllUsers
      tags: [Users]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageUserDto'
    post:
      summary: Create a new user
      operationId: createUser
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'

  /api/users/{id}:
    get:
      summary: Get user by ID
      operationId: getUserById
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: User not found

    delete:
      summary: Delete user
      operationId: deleteUser
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted

  # CALCULATORS
  /api/calculators:
    get:
      summary: Get all calculators
      operationId: getAllCalculators
      tags: [Calculators]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of calculators
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageCalculatorDto'
    post:
      summary: Create a new calculator
      operationId: createCalculator
      tags: [Calculators]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCalculatorRequest'
      responses:
        '201':
          description: Calculator created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculatorDto'

  /api/calculators/{id}:
    get:
      summary: Get calculator by ID
      operationId: getCalculatorById
      tags: [Calculators]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Calculator found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculatorDto'

  /api/calculators/{id}/versions:
    get:
      summary: Get calculator versions
      operationId: getCalculatorVersions
      tags: [Calculators]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CalculatorVersionDto'
    post:
      summary: Add a new version
      operationId: addCalculatorVersion
      tags: [Calculators]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVersionRequest'
      responses:
        '201':
          description: Version added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculatorVersionDto'

  # CALCULATIONS
  /api/calculations:
    get:
      summary: Get user calculations
      operationId: getCalculations
      tags: [Calculations]
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of calculations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageCalculationDto'
    post:
      summary: Create a new calculation
      operationId: createCalculation
      tags: [Calculations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCalculationRequest'
      responses:
        '201':
          description: Calculation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationDto'

  /api/calculations/{id}/scenarios:
    get:
      summary: Get scenarios for calculation
      operationId: getScenarios
      tags: [Calculations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of scenarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CalculationScenarioDto'
    post:
      summary: Add a scenario
      operationId: addScenario
      tags: [Calculations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScenarioRequest'
      responses:
        '201':
          description: Scenario added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationScenarioDto'

  /api/calculations/{id}/select-scenario/{scenarioId}:
    put:
      summary: Select active scenario
      operationId: selectScenario
      tags: [Calculations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scenarioId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Scenario selected

  # PORTFOLIO
  /api/portfolio:
    get:
      summary: Get current user's portfolio (cash flows from selected scenarios)
      operationId: getPortfolio
      tags: [Portfolio]
      parameters:
        - name: from
          in: query
          required: false
          schema: { type: string, format: date }
          description: Inclusive start date. Defaults to first day of current month.
        - name: to
          in: query
          required: false
          schema: { type: string, format: date }
          description: Inclusive end date. Defaults to last day of current month.
      responses:
        '200':
          description: Portfolio for current user, grouped by currency
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioDto' }

  /api/portfolio/items/{calculationId}:
    delete:
      summary: Remove a calculation from current user's portfolio
      operationId: removeFromPortfolio
      tags: [Portfolio]
      parameters:
        - name: calculationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Removed
        '404':
          description: Item not found in current user's portfolio

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
  schemas:
    # AUTH
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    ProductType:
      type: string
      enum: [LOAN, DEPOSIT]
    
    InterestMethod:
      type: string
      enum: [SIMPLE, COMPOUND]
    
    DayCountConvention:
      type: string
      enum: [ACTUAL_365, ACTUAL_360, THIRTY_360, ACTUAL_ACTUAL]
    
    RateType:
      type: string
      enum: [FIXED, FLOATING]
    
    Frequency:
      type: string
      enum: [ONCE, DAILY, WEEKLY, BI_WEEKLY, MONTHLY, QUARTERLY, SEMI_ANNUALLY, ANNUALLY]
    
    AmortizationType:
      type: string
      enum: [ANNUITY, LINEAR, BULLET, ZERO_COUPON]
    
    CurrencyCode:
      type: string
      enum: [USD, EUR, UAH, GBP]
    
    RoundingMode:
      type: string
      enum: [UP, DOWN, CEILING, FLOOR, HALF_UP, HALF_DOWN, HALF_EVEN, UNNECESSARY]

    UserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        profilePictureUrl:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time

    CreateUserRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
        password:
          type: string
          minLength: 8
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        profilePictureUrl:
          type: string
        role:
          type: string

    PageUserDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserDto'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer

    CalculatorDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time

    CreateCalculatorRequest:
      type: object
      required: [code, name, algorithmMetadata]
      properties:
        code:
          type: string
          pattern: '^[a-zA-Z0-9_-]+$'
        name:
          type: string
        description:
          type: string
        algorithmMetadata:
          $ref: '#/components/schemas/FinancialProductDefinitionDto'

    CalculatorVersionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        calculatorId:
          type: string
          format: uuid
        version:
          type: integer
        algorithmMetadata:
          $ref: '#/components/schemas/FinancialProductDefinitionDto'
        createdAt:
          type: string
          format: date-time

    CreateVersionRequest:
      type: object
      required: [algorithmMetadata]
      properties:
        algorithmMetadata:
          $ref: '#/components/schemas/FinancialProductDefinitionDto'

    FinancialProductDefinitionDto:
      type: object
      required: [type, interest, repayment]
      properties:
        type:
          $ref: '#/components/schemas/ProductType'
        interest:
          $ref: '#/components/schemas/InterestConfigDto'
        repayment:
          $ref: '#/components/schemas/RepaymentConfigDto'
        constraints:
          $ref: '#/components/schemas/ProductConstraintsDto'
        defaults:
          $ref: '#/components/schemas/ProductDefaultsDto'

    InterestConfigDto:
      type: object
      required: [method, dayCountConvention, rateType]
      properties:
        method:
          $ref: '#/components/schemas/InterestMethod'
        dayCountConvention:
          $ref: '#/components/schemas/DayCountConvention'
        rateType:
          $ref: '#/components/schemas/RateType'
        accrualFrequency:
          $ref: '#/components/schemas/Frequency'
        compoundingFrequency:
          $ref: '#/components/schemas/Frequency'

    RepaymentConfigDto:
      type: object
      required: [strategy, frequency]
      properties:
        strategy:
          $ref: '#/components/schemas/AmortizationType'
        frequency:
          $ref: '#/components/schemas/Frequency'

    ProductConstraintsDto:
      type: object
      properties:
        minAmount:
          type: number
        maxAmount:
          type: number
        minTerm:
          type: integer
        maxTerm:
          type: integer
        minRate:
          type: number
        maxRate:
          type: number

    ProductDefaultsDto:
      type: object
      properties:
        fixedRate:
          type: number
        currency:
          $ref: '#/components/schemas/CurrencyCode'
          default: "USD"
        roundingScale:
          type: integer
          default: 2
        roundingMode:
          $ref: '#/components/schemas/RoundingMode'
          default: "HALF_UP"

    CalculationInputDto:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
        rate:
          type: number
        term:
          type: integer
        startDate:
          type: string
          format: date

    PageCalculatorDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CalculatorDto'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer

    CalculationDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        calculatorVersionId:
          type: string
          format: uuid
        currency:
          type: string
        selectedScenarioId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time

    CreateCalculationRequest:
      type: object
      required: [userId, calculatorId, inputData]
      properties:
        userId:
          type: string
          format: uuid
        calculatorId:
          type: string
          format: uuid
        inputData:
          $ref: '#/components/schemas/CalculationInputDto'
        currency:
          type: string
          minLength: 3
          maxLength: 3

    CalculationScenarioDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        calculationId:
          type: string
          format: uuid
        scenarioName:
          type: string
        scenarioInput:
          $ref: '#/components/schemas/CalculationInputDto'
        scenarioResult:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time

    CreateScenarioRequest:
      type: object
      required: [scenarioInput]
      properties:
        scenarioName:
          type: string
        scenarioInput:
          $ref: '#/components/schemas/CalculationInputDto'

    PageCalculationDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CalculationDto'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer

    PortfolioDto:
      type: object
      properties:
        userId:  { type: string, format: uuid }
        from:    { type: string, format: date }
        to:      { type: string, format: date }
        buckets:
          type: array
          items: { $ref: '#/components/schemas/PortfolioCurrencyBucketDto' }

    PortfolioCurrencyBucketDto:
      type: object
      properties:
        currency:     { type: string }
        totalInflow:  { type: number }
        totalOutflow: { type: number }
        netCashFlow:  { type: number }
        items:
          type: array
          items: { $ref: '#/components/schemas/PortfolioItemDto' }

    PortfolioItemDto:
      type: object
      properties:
        date:           { type: string, format: date }
        type:
          type: string
          enum: [INFLOW, OUTFLOW]
        amount:         { type: number }
        currency:       { type: string }
        description:    { type: string }
        principal:      { type: number }
        interest:       { type: number }
        fee:            { type: number }
        calculationId:  { type: string, format: uuid }
        scenarioId:     { type: string, format: uuid }